updated game control to be individual folder and combined game.py to gamestate.py
