+++ src/TeamControl/GameControl/Processing.py
@@ -1,0 +1,6 @@
+from TeamControl.GameControl.GameState import *
+
+class Processing():
+    def __init__(self):
+        ...
+
